app/requirements.txt
django>=5,<6
djangorestframework>=3.10
mistune>=2.0.3 # not directly required, pinned by Snyk to avoid a vulnerability
drf-spectacular # for schema generation
uritemplate  # for python manage.py generateschema
django-compressor==4.4
psycopg2-binary

-e file:../python_package
